research_code/requirements.txt
einops==0.7.0
matplotlib==3.7.2
nibabel==5.1.0
numpy==1.24.4
pandas==2.1.4
Pillow==10.1.0
prettytable==3.9.0
scikit_learn==1.3.0
scipy==1.10.1
torch==2.1.0
torchvision==0.16.0
# nppy==0.2.1
